pyproject.toml
[project]
name = "cicd-helloworld-example"
version = "0.0.1"
description = "This is n project used as an example in the CI/CD flow at meemoo"
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    "viaa-chassis",
    "litestar[standard]>=2.19.0,<3.0.0"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.12",
]
[project.optional-dependencies]
dev = [
    "ruff",
    "pytest>=9.0.2,<10.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--rootdir=./app/ -v"
testpaths = ["tests"]
